Several constants and maps have been moved from wfexs_backend.ro_crate module to `wfexs_backend.utils.rocrate` one
